+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - support update changelog automatically. (2024-2-14)
 - Add chinese document. (2024-2-15)
 - Update document. (2024-2-15)
+- Support report page performance data. (2024-3-10)